# Improve progress performance