--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/DebugInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/DebugInternal.cs
@@ -340,15 +340,6 @@ namespace Proto.Promises
                         }
                     }
                 }
-                
-                new protected void Dispose()
-                {
-                    base.Dispose();
-                    lock (_previousPromises)
-                    {
-                        _previousPromises.Clear();
-                    }
-                }
             }
         }
 #else // PROMISE_DEBUG