Resolves #115

Refactored progress to iterate over the promise tree when progress is subscribed instead of when progress is reported. This moves the heavy work to the 1-time hookup rather than every time progress is reported. Reporting progress is now much faster and scales better with longer promise chains (`O(n)` where `n` is now Merge/Race/Preserve promises instead of every promise). This also increases memory when progress is listened to, but decreases it when it is not (for `.Then` chains).

Progress now uses full 32-bit float precision instead of 16-bit fixed precision.

---

Without listening to progress or normalizing with `.AwaitWithProgress`. (8 awaits with max depth 3)

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 2.184 us |         - |    696 B |
| ContinueWith |    True | 2.594 us |         - |    400 B |

This PR:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 2.075 us |         - |    696 B |
| ContinueWith |    True | 2.104 us |         - |    384 B |

---

Same setup, but with listening to progress, normalizing with `.AwaitWithProgress`, and reporting `0.5` progress before resolving.

Master:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 5.138 us |         - |    896 B |
| ContinueWith |    True | 4.691 us |         - |    568 B |

This PR:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 4.548 us |         - |    936 B |
| ContinueWith |    True | 4.222 us |         - |    592 B |

Longer promise chains would result in even larger speed differences due to the better scaling.

---

Same setup, but with normalizing with `.AwaitWithProgress`, and reporting `0.5` progress before resolving, and _without_ listening to progress.

Master:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 3.885 us |         - |    728 B |
| ContinueWith |    True | 3.516 us |         - |    400 B |

This PR:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 2.587 us |         - |    728 B |
| ContinueWith |    True | 2.270 us |         - |    384 B |

In this PR, reporting progress when there is no progress listener is an `O(1)` no-op. In master it has to iterate over the entire promise tree, so larger promise trees will take more time to report the progress, even without a listener.